Feature Added: Confirmation Dialog message Before delete item, default Message when fragment is empty

+++ app/src/main/java/com/network/tafasir/UI/Fragments/BookMarkF.kt
@@ -1,11 +1,13 @@
 package com.network.tafasir.UI.Fragments
 
+import android.app.AlertDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.Adapter
+import android.widget.TextView
 import androidx.fragment.app.activityViewModels
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
@@ -26,7 +28,10 @@ import kotlinx.coroutines.launch
 
 class BookMarkF : Fragment() {
 
-    private val bookMarkViewModel : BookMarkViewModel by activityViewModels()
+    private val bookMarkViewModel: BookMarkViewModel by activityViewModels()
+    private var isListEmpty = true
+    private lateinit var bookMarkDefaultMessageTv: TextView
+    private lateinit var recyclerView: RecyclerView
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
@@ -37,25 +42,60 @@ class BookMarkF : Fragment() {
         val paddingControl = PaddingControl(view)
         var adapter: BookmarkAdapter? = null
 
-        val recyclerView: RecyclerView = view.findViewById(R.id.bookmarkRecycler)
+        bookMarkDefaultMessageTv = view.findViewById<TextView>(R.id.bookMarkDefaultMessageTv)
+        recyclerView = view.findViewById(R.id.bookmarkRecycler)
 
         bookMarkViewModel.getAll() // Get Data
-        bookMarkViewModel.result.observe(viewLifecycleOwner, Observer {list ->
+        bookMarkViewModel.result.observe(viewLifecycleOwner, Observer { list ->
             //
-                adapter = BookmarkAdapter(context, list, { position ->
+
+            if (!list.isEmpty()) {
+                isListEmpty = false
+                hideDefaultMessage()
+            }
+
+            adapter = BookmarkAdapter(context, list, { position ->
                 openAyah(position, list)
 
 
-            }) { position ->
+            }, { position ->
 
-                    bookMarkViewModel.delete(list[position])
-                    list.removeAt(position)
-                    adapter?.notifyDataSetChanged()
+                //Alert Dialog To Confirm Deleting Selected BookMark
+
+                val builder = AlertDialog.Builder(context)
+
+                builder.setMessage(getString(R.string.areYouSureFromDeleteAction))
+                    .setCancelable(false)
+                    .setPositiveButton(getString(R.string.yes)) { dialog, id ->
+
+                        // Delete selected
+                        bookMarkViewModel.delete(list[position])
+                        list.removeAt(position)
+                        adapter?.notifyDataSetChanged()
+
+                    }
+                    .setNegativeButton(getString(R.string.no)) { dialog, id ->
+                        // Dismiss the dialog
+                        dialog.dismiss()
+                    }
+
+                val alert = builder.create()
+                alert.show()
+            }, {
+                //Empty Adapter Listener
+                if(list.isEmpty()){
+                    isListEmpty = true
+                    hideDefaultMessage()
+                }else{
+                    isListEmpty = false
+                    hideDefaultMessage()
                 }
+            })
 
 
 
-            recyclerView.layoutManager = LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
+            recyclerView.layoutManager =
+                LinearLayoutManager(context, LinearLayoutManager.VERTICAL, false)
             recyclerView.adapter = adapter
 
             //Padding Control
@@ -85,6 +125,17 @@ class BookMarkF : Fragment() {
     private fun openAyah(posision: Int, list: List<BookMarkEntity>?) {
         val transaction = activity as Communicat
         transaction.TransferData(list!![posision].soraNumber, 1, list!![posision].ayahNumber)
+    }
+
+    private fun hideDefaultMessage() {
+
+        if (isListEmpty) {
+            recyclerView.visibility = View.GONE
+            bookMarkDefaultMessageTv.visibility = View.VISIBLE
+        } else {
+            recyclerView.visibility = View.VISIBLE
+            bookMarkDefaultMessageTv.visibility = View.GONE
+        }
     }
 
 }
