+++ app/src/main/java/com/network/tafasir/UI/Fragments/FavoriteF.kt
@@ -1,25 +1,21 @@
 package com.network.tafasir.UI.Fragments
 
-import android.content.ClipData
-import android.content.ClipboardManager
-import android.content.Context
-import android.content.Intent
+import android.app.AlertDialog
 import android.os.Bundle
+import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.PopupMenu
+import android.widget.TextView
 import android.widget.Toast
-import androidx.core.content.ContextCompat
+import androidx.core.view.size
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.activityViewModels
-import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.RecyclerView.OnScrollListener
-import com.network.tafasir.DATA.Database.DataClasses.SoraWithTafsir
-import com.network.tafasir.DATA.Database.Room.Favorite.FavoriteEntity
 import com.network.tafasir.R
 import com.network.tafasir.UI.Adapters.Recycler.FavoriteAdapter
 import com.network.tafasir.UI.Controlers.ShareContent
@@ -31,11 +27,15 @@ class FavoriteF : Fragment() {
 
     private val favoriteViewModel: FavoriteViewModel by activityViewModels()
     private val mainViewModel: MainViewModel by activityViewModels()
+    private lateinit var recyclerView: RecyclerView
     private var adapter: FavoriteAdapter? = null
-    private lateinit var tfsir:String
-    private lateinit var mofsirName:String
+    private lateinit var tfsir: String
+    private lateinit var mofsirName: String
     private var shareIt = ShareContent()
     private var isDefault = true
+    private var isListEmpty = true
+    private lateinit var favoriteDefaultMessageTv: TextView
+    val TAG_F = "Favorite"
 
 
     override fun onCreateView(
@@ -45,52 +45,89 @@ class FavoriteF : Fragment() {
 
         val view = inflater.inflate(R.layout.fragment_favorite, container, false)
         val context = view.context
-
+        favoriteDefaultMessageTv = view.findViewById<TextView>(R.id.favoriteDefaultMessageTv)
 
         val paddingControl = PaddingControl(view)
 
-        val recyclerView:RecyclerView = view.findViewById(R.id.favoriteRecycler)
+        recyclerView = view.findViewById(R.id.favoriteRecycler)
 
         favoriteViewModel.getAll()
         favoriteViewModel.allFavorite.observe(viewLifecycleOwner, Observer { list ->
             //
 
+            if(!list.isEmpty()){
+                isListEmpty = false
+                hideDefaultMessage()
+            }
 
 
-            adapter = FavoriteAdapter(context, list
+            adapter = FavoriteAdapter(context, list, { position ->
 
-                ,{ position ->
+                //Alert Dialog To Confirm Deleting Selected Saved Ayah
 
-                    //Delete
-                    favoriteViewModel.delete(list[position])
-                    list.removeAt(position)
-                    adapter?.notifyDataSetChanged()
+                val builder = AlertDialog.Builder(context)
 
-                },{ position, view ->
+                builder.setMessage(getString(R.string.areYouSureFromDeleteAction))
+                    .setCancelable(false)
+                    .setPositiveButton(getString(R.string.yes)) { dialog, id ->
 
-                  //Tafsir
-                    tafsirPopUpMenu(view, list[position].sora_number, list[position].ayah_number, position)
+                        // Delete selected
+                        favoriteViewModel.delete(list[position])
+                        list.removeAt(position)
+                        adapter?.notifyDataSetChanged()
 
-                },{ position, view ->
-
-                    //Share section
-
-                    //set Default Data
-                    if (isDefault){
-                        mofsirName = getString(R.string.moyasar)
-                        tfsir = list[position].tafsir
+                    }
+                    .setNegativeButton(getString(R.string.no)) { dialog, id ->
+                        // Dismiss the dialog
+                        dialog.dismiss()
                     }
 
-                    val data = ShareContent.SharedData(list[position].soraName,
-                        list[position].ayah,
-                        list[position].sora_number,
-                        list[position].ayah_number,
-                        tfsir,
-                        mofsirName)
+                val alert = builder.create()
+                alert.show()
 
-                    shareIt.shareContent(data, view, context)
+            }, { position, view ->
 
+                //Tafsir
+                tafsirPopUpMenu(
+                    view,
+                    list[position].sora_number,
+                    list[position].ayah_number,
+                    position
+                )
+
+            }, { position, view ->
+
+                //Share section
+
+                //set Default Data
+                if (isDefault) {
+                    mofsirName = getString(R.string.moyasar)
+                    tfsir = list[position].tafsir
+                }
+
+                val data = ShareContent.SharedData(
+                    list[position].soraName,
+                    list[position].ayah,
+                    list[position].sora_number,
+                    list[position].ayah_number,
+                    tfsir,
+                    mofsirName
+                )
+
+                shareIt.shareContent(data, view, context)
+
+            }, {
+                //Empty Adapter Listener
+                if(list.isEmpty()){
+                    isListEmpty = true
+                    hideDefaultMessage()
+                }else{
+                    isListEmpty = false
+                    hideDefaultMessage()
+                }
             })
+
+
             recyclerView.layoutManager = LinearLayoutManager(
                 context,
                 LinearLayoutManager.VERTICAL,
@@ -121,7 +158,7 @@ class FavoriteF : Fragment() {
         return view
     }
 
-    private fun tafsirPopUpMenu(view: View, soraNumber: Int, ayahNumber: Int, position: Int)     {
+    private fun tafsirPopUpMenu(view: View, soraNumber: Int, ayahNumber: Int, position: Int) {
 
         val popup = PopupMenu(context, view)
         popup.inflate(R.menu.tafasir_menu)
@@ -152,7 +189,8 @@ class FavoriteF : Fragment() {
                 R.id.ibenKatir -> {
 
                     changeTafsir(soraNumber, ayahNumber, 4, position)
-                    Toast.makeText(context, getString(R.string.ibenKatir), Toast.LENGTH_SHORT).show()
+                    Toast.makeText(context, getString(R.string.ibenKatir), Toast.LENGTH_SHORT)
+                        .show()
 
                 }
                 R.id.tentawi -> {
@@ -189,9 +227,9 @@ class FavoriteF : Fragment() {
 
     }
 
-    private fun changeTafsir(soraNum:Int, ayahNum:Int, tafsirNum:Int, position:Int){
+    private fun changeTafsir(soraNum: Int, ayahNum: Int, tafsirNum: Int, position: Int) {
 
-        mainViewModel.getTafsir(soraNum, ayahNum, tafsirNum){it ->
+        mainViewModel.getTafsir(soraNum, ayahNum, tafsirNum) { it ->
             tfsir = it.tafsir
             mofsirName = it.mofasir_name
             adapter!!.changeTafsir(it.tafsir, position)
@@ -200,6 +238,15 @@ class FavoriteF : Fragment() {
     }
 
 
+    private fun hideDefaultMessage() {
 
+        if (isListEmpty) {
+            recyclerView.visibility = View.GONE
+            favoriteDefaultMessageTv.visibility = View.VISIBLE
+        } else {
+            recyclerView.visibility = View.VISIBLE
+            favoriteDefaultMessageTv.visibility = View.GONE
+        }
+    }
 
 }
